[BasePawn][CarPawn][CarAIController] Made AI system for a CarPawn - Fixed Switch lanes in base classes and CarPawn class - Added AI logic to a CarPawn controller - Made a randomizer for CarPawn initial speed and accel values - Finished CarPawn methods for auto-checking Nearby lanes for other cars  - Deleted empty Behaviour tree and Blackboard

+++ ue4/Source/SkateTraffic/Components/SkaterPawnMovementComponent.cpp
@@ -54,7 +54,7 @@ void USkaterPawnMovementComponent::PushForwardEnd()
 	CurrentPushForwardVelocity = 0.f;
 }
 
-void USkaterPawnMovementComponent::SwitchLaneStart(int8 Direction)
+void USkaterPawnMovementComponent::SwitchLaneStart(int32 Direction)
 {
 	if (PawnOwner->IsPushingForward())
 	{
